model deployment/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
transformers
torch 
pydantic
Pillow
requests
